Adjust the login and register of salon and user.
Add the user name and the button of logout when the user is logged in and Add the the button of login and register when the user is not logged in at the top of meifan.
Adjust the html and css of login and register in a view, use the tab to distinguish.